validation

+++ Services/DTO/ChargeDTO.cs
@@ -5,7 +5,7 @@
         public long Id { get; set; }
         public decimal Days { get; set; }
         public decimal Rate { get; set; }
-        public decimal Amount => this.Days * this.Rate;
+        public decimal Amount => (decimal)this.Days * (decimal)this.Rate;
         public long LookupId { get; set; }
     }
 }
